update to use new helper code
